+++ src/controllers/enrollment.controller.ts
@@ -6,7 +6,7 @@ export namespace EnrollmentController {
 		try {
 			const { userId, courseId } = req.body;
 			if (!userId || !courseId) {
-				throw new Error("Please provide both user and course id in req body");
+				throw new Error("Please provide both user and course id");
 			}
 			const enrollment = await EnrollmentService.create({ userId, courseId });
 			return res.status(201).json({
@@ -20,8 +20,11 @@ export namespace EnrollmentController {
 
 	export const findByUserId = async (req: Request, res: Response, next: NextFunction) => {
 		try {
-			const userId = req.params.id;
-			const enrollments = await EnrollmentService.findByUserId(userId);
+			const id = req.params.id;
+			if (!id) {
+				return next(new Error("Please provide User id"));
+			}
+			const enrollments = await EnrollmentService.findByUserId(id);
 			return res.status(200).json({ data: enrollments });
 		} catch (error: any) {
 			throw new Error(error);
@@ -34,8 +37,11 @@ export namespace EnrollmentController {
 		next: NextFunction
 	) => {
 		try {
-			const courseId = req.params.id;
-			const enrollments = await EnrollmentService.findByCourseId(courseId);
+			const id = req.params.id;
+			if (!id) {
+				return next(new Error("Please provide Course id"));
+			}
+			const enrollments = await EnrollmentService.findByCourseId(id);
 			return res.status(200).json({ data: enrollments });
 		} catch (error: any) {
 			throw new Error(error);
@@ -45,6 +51,9 @@ export namespace EnrollmentController {
 	export const deleteById = async (req: Request, res: Response, next: NextFunction) => {
 		try {
 			const { userId, courseId } = req.params;
+			if (!userId || !courseId) {
+				return next(new Error("Please provide Lesson id, course id"));
+			}
 			await EnrollmentService.deleteById(userId, courseId);
 			return res.status(204).json({ message: "Enrollment cancelled successfully" });
 		} catch (error: any) {
